requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
colorama==0.4.6
idna==3.4
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
shodan==1.28.0
urllib3==1.26.14
XlsxWriter==3.0.8
